// src/app/api/shared/task/task-default-assignee-response.interface.ts
export interface TaskDefaultAssigneeResponseInterface extends Array<{
  agency_id: string;
  company: string;
  contact_id: string;
  email: string;
  firstname: string;
  firstname2: string;
  is_main: string;
  language_id: string;
  lastname: string;
  lastname2: string;
  search: string;
}> {

}